When FileSystemState.changeFile() release data from unlinkedUnitStore.
This fixes b/269188397 in Cider because now unlikned data is present
in both ByteStore and UnlinkedUnitStore, or absent in both. Before
this change we would get it from UnlinkedUnitStore, but expected
that these bytes were also requested from ByteStore, so we attempted
to release them (and got "underflow" of reference counter).

Also, we now put the unlinked data into UnlinkedUnitStore not only
after reading bytes, but also after building it. So, we second use
will get it from the store, not the third.

Change-Id: I89bf554511d3be82f67ef1f237061ba8a4693592
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/284745
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>