[analysis_server] Fix ordering of edits that include multiple inserts at the same location
While testing "Move to File" I found the file header was being inserted at the bottom of the document in VS Code, despite the tests working correctly.

This turned out to be because we were sending edits in the same order as server generates (which are designed to be applied sequentially, so they are essentially reversed) on the assumption that order doesn't matter to LSP. This is not correct for inserts with the same offset - those are expected to appear in-order.

Because our tests also just applied the edits in reverse-order, they passed but didn't match the behaviour in VS Code.

This change reverses the order of edits when mapping to LSP (so same-offset inserts are back in "forwards order") and also updates the tests to re-sort in an order that matches VS Code's behaviour.

Change-Id: I40d6be6899e4f13dbff52f39fefb3bbd2caa3521
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/285500
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>