[_fe_analyzer_shared] Support unions in exhaustiveness algorithm
This adds support for union directly in the exhaustiveness algorithm
meaning that we can support deeply nested or-patterns without having
to eagerly expand the patterns.

For instance, without support for unions,

  case (true || false, true || false):

would have to be expanded to

  case (true, true):
  case (true, false):
  case (false, true):
  case (false, false):

before performing exhaustiveness.

Now, we instead perform the expansion in-place and only we need to.
That is, when filtering by subtype `true` of `bool`, we only consider
the row

  (true, true || false)

as remaining, because the row

  (false, true || false)

won't match anyway.

Change-Id: I1a6402bd21b81bf40924ac0048b683936ce9f608
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/285263
Commit-Queue: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>