[stable][vm, gc] Add missing promotion of Finalizer external size. Remove race incrementing external size.
When a FinalizerEntry's target gets promoted, the associated external size needs to also get promoted. We were handling the cases where the FinalizerEntry itself was either already old or remained new, but not the case where it was promoted. Failing to promote the external size meant that when the finalizer was collected, external size was subtraced from old space that was still being attributed to new space, so the old space total external size became negative.

TEST=ci
Bug: https://github.com/dart-lang/sdk/issues/50537
Change-Id: I83316636dea13415f38343212157c32f5ef19857
Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/272350
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/289941
Reviewed-by: Alexander Markov <alexmarkov@google.com>